Change signature of Add function

+++ Cpp for Cs Developers Practice/Maths/math.cpp
@@ -1,6 +1,6 @@
 #include "math.hpp"
 
-int Add(int a, int b)
+float Add(float a, float b)
 {
 	return a + b;
 }
